Move incident factories to right place
.. and tune annotations some more.
