+++ src/argus/dev/management/commands/check_token_expiry.py
@@ -6,8 +6,8 @@ from django.db.models import Q
 from django.utils import timezone
 from rest_framework.authtoken.models import Token
 
+from argus.incident.factories import create_token_expiry_incident
 from argus.incident.models import (
-    create_token_expiry_incident,
     get_or_create_default_instances,
     Incident,
     Tag,
