+++ tests/dev/test_check_token_expiry.py
@@ -13,8 +13,8 @@ from argus.dev.management.commands.check_token_expiry import (
     get_tokens_without_expiry_incident,
 )
 from argus.auth.factories import SourceUserFactory
-from argus.incident.factories import SourceSystemFactory
-from argus.incident.models import Incident, Tag, create_token_expiry_incident
+from argus.incident.factories import SourceSystemFactory, create_token_expiry_incident
+from argus.incident.models import Incident, Tag
 from argus.util.testing import connect_signals, disconnect_signals
