+++ tests/notificationprofile/test_github236.py
@@ -3,7 +3,8 @@ from django.test import TestCase, tag
 
 from argus.auth.factories import PersonUserFactory
 from argus.filter.factories import FilterFactory
-from argus.incident.models import create_fake_incident, get_or_create_default_instances, Event
+from argus.incident.models import get_or_create_default_instances, Event
+from argus.incident.factories import create_fake_incident
 from argus.notificationprofile import factories
 from argus.notificationprofile.media import send_notifications_to_users
 from argus.util.testing import disconnect_signals, connect_signals
